Split exiftool reader tests into their own files
Move ExifToolRead tests to tests/unit/test_exiftool_read.py and
ExifToolReadVideo camera-uuid tests to tests/unit/test_exiftool_read_video.py
so that test_exifread.py only covers exif_read.
